# projects/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "projects-service dev",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/object_status/list": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Get Object Statuses",
        "description": "Gets a paginated list of object statuses",
        "operationId": "get_object_statuses_api_object_status_list_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_IObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/object_status/{id}": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Read Object Status By Id",
        "description": "Gets an object status by id",
        "operationId": "read_object_status_by_id_api_object_status__id__get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Update Object Status",
        "description": "Updates an object status by its id",
        "operationId": "update_object_status_api_object_status__id__put",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IObjectStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPutResponseBase_IObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Delete Object Status",
        "description": "Deletes an object status by its id",
        "operationId": "delete_object_status_api_object_status__id__delete",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IDeleteResponseBase_IObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/object_status/{id}/objects": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Get Object Status With Objects List By Id",
        "description": "Gets object status with all objects list by id",
        "operationId": "get_object_status_with_objects_list_by_id_api_object_status__id__objects_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IObjectStatusWithAllObjectsListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/object_status/{id}/towers": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Get Object Status With Towers List By Id",
        "description": "Gets object status with towers list by id",
        "operationId": "get_object_status_with_towers_list_by_id_api_object_status__id__towers_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IObjectStatusWithTowersListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/object_status/{id}/spans": {
      "get": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Get Object Status With Spans List By Id",
        "description": "Gets object status with spans list by id",
        "operationId": "get_object_status_with_spans_list_by_id_api_object_status__id__spans_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IObjectStatusWithSpansListRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/object_status": {
      "post": {
        "tags": [
          "ObjectStatus"
        ],
        "summary": "Create Object Status",
        "description": "Creates a new object status",
        "operationId": "create_object_status_api_object_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IObjectStatusCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPostResponseBase_IObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project/list/status": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Projects List By Project Status Id",
        "description": "Gets a paginated list of projects by status id",
        "operationId": "get_projects_list_by_project_status_id_api_project_list_status_get",
        "parameters": [
          {
            "description": "The UUID id of the project status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project status",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_IProjectRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project/{id}/status": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project With Status By Id",
        "description": "Gets a project by id",
        "operationId": "get_project_with_status_by_id_api_project__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the project",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the project",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_IProjectWithStatus_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subproject/{id}/status": {
      "get": {
        "tags": [
          "Subprojects"
        ],
        "summary": "Get Subproject With Status By Id",
        "description": "Gets a subproject with status by id",
        "operationId": "get_subproject_with_status_by_id_api_subproject__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the subproject",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the subproject",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ISubprojectWithStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/list/status": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Get Towers List By Status Id",
        "description": "Gets a paginated list of towers by status id",
        "operationId": "get_towers_list_by_status_id_api_tower_list_status_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_ITowerRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tower/{id}/status": {
      "get": {
        "tags": [
          "Towers"
        ],
        "summary": "Get Tower With Object Status By Id",
        "description": "Gets a tower with status by id",
        "operationId": "get_tower_with_object_status_by_id_api_tower__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the tower",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the tower",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ITowerWithObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/span/list/status": {
      "get": {
        "tags": [
          "Spans"
        ],
        "summary": "Get Spans List By Status Id",
        "description": "Gets a paginated list of spans by status id",
        "operationId": "get_spans_list_by_status_id_api_span_list_status_get",
        "parameters": [
          {
            "description": "The UUID id of the object status",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the object status",
              "format": "uuid"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1.0,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponsePaginated_ISpanRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/span/{id}/status": {
      "get": {
        "tags": [
          "Spans"
        ],
        "summary": "Get Span With Object Status By Id",
        "description": "Gets a span with status by id",
        "operationId": "get_span_with_object_status_by_id_api_span__id__status_get",
        "parameters": [
          {
            "description": "The UUID id of the span",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string",
              "description": "The UUID id of the span",
              "format": "uuid"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IGetResponseBase_ISpanWithObjectStatusRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "IDeleteResponseBase_IObjectStatusRead_": {
        "title": "IDeleteResponseBase[IObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data deleted correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusRead"
          }
        }
      },
...
      "IGetResponseBase_IObjectStatusRead_": {
        "title": "IGetResponseBase[IObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusRead"
          }
        }
      },
      "IGetResponseBase_IObjectStatusWithAllObjectsListRead_": {
        "title": "IGetResponseBase[IObjectStatusWithAllObjectsListRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusWithAllObjectsListRead"
          }
        }
      },
      "IGetResponseBase_IObjectStatusWithSpansListRead_": {
        "title": "IGetResponseBase[IObjectStatusWithSpansListRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusWithSpansListRead"
          }
        }
      },
      "IGetResponseBase_IObjectStatusWithTowersListRead_": {
        "title": "IGetResponseBase[IObjectStatusWithTowersListRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusWithTowersListRead"
          }
        }
      },
...
      "IGetResponseBase_IProjectWithStatus_": {
        "title": "IGetResponseBase[IProjectWithStatus]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IProjectWithStatus"
          }
        }
      },
...
      "IGetResponseBase_ISpanWithObjectStatusRead_": {
        "title": "IGetResponseBase[ISpanWithObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ISpanWithObjectStatusRead"
          }
        }
      },
...
      "IGetResponseBase_ISubprojectWithStatusRead_": {
        "title": "IGetResponseBase[ISubprojectWithStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ISubprojectWithStatusRead"
          }
        }
      },
...
      "IGetResponseBase_ITowerWithObjectStatusRead_": {
        "title": "IGetResponseBase[ITowerWithObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/ITowerWithObjectStatusRead"
          }
        }
      },
...
      "IGetResponsePaginated_IObjectStatusRead_": {
        "title": "IGetResponsePaginated[IObjectStatusRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__IObjectStatusRead_"
          }
        }
      },
      "IGetResponsePaginated_IProjectRead_": {
        "title": "IGetResponsePaginated[IProjectRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__IProjectRead_"
          }
        }
      },
...
      "IGetResponsePaginated_ISpanRead_": {
        "title": "IGetResponsePaginated[ISpanRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__ISpanRead_"
          }
        }
      },
...
      "IGetResponsePaginated_ITowerRead_": {
        "title": "IGetResponsePaginated[ITowerRead]",
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data got correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/PageBase_DataType__ITowerRead_"
          }
        }
      },
...
      "IObjectStatusCreate": {
        "title": "IObjectStatusCreate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "IObjectStatusUpdate": {
        "title": "IObjectStatusUpdate",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "IPostResponseBase_IObjectStatusRead_": {
        "title": "IPostResponseBase[IObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data created correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusRead"
          }
        }
      },
...
      "IPutResponseBase_IObjectStatusRead_": {
        "title": "IPutResponseBase[IObjectStatusRead]",
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string",
            "default": "Data updated correctly"
          },
          "meta": {
            "title": "Meta",
            "type": "object",
            "default": {}
          },
          "data": {
            "$ref": "#/components/schemas/IObjectStatusRead"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    }
  }
}